ATTEMPT AT EMBEDING BOOKING WIDGET - not complete
Problem lies in the height of the widget not resizing as the window
grow. This typically happens after you pick dates and the number of guests
that will stay. The quote automatically appears and the content of the
window shifts down and out of the top level view.
I've considered that we could use the postMessage function to communicate
between the widget window and the parent but, it dones't seem viable
considering I don't have control over what messages the widget can send.