fix topImg bleeding into footer post inquiry submission
Experienced issue in landscape view of small laptop(1024), upon
submitting inquiry on booking page, the top img was showing on top of the footer but under the booking widget.

THE PROBLEM stemmed from the footer which did not have any padding,
only margin. This meant there was no footer background for the top img
 to hide behind.

THE SOLUTION was to simply replace margin with padding for component
to take up room and keep topImg from showing.